backend/requirements.txt
Django>=4.2
psycopg2-binary
django-cors-headers
Pillow
djangorestframework>=3.15.0